+++ Changes.md
@@ -1,5 +1,12 @@
 # Changes in Emacs MMA mode
 
+## 2018-12-29
+
+* Add support for xplayer MIDI player and make this player default.
+
+* Add `//` comment syntax for use with `comment-dwim` and
+  `comment-region`.
+  
 ## 2014-02-08
 
 * Emacs-23.4.1 version.